requirements.txt
# ── Core Django ─────────────────────────────────────────────────
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.3.1
django-celery-results==2.5.1
dj-database-url==2.1.0
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3

# ── Async / Task Queue ───────────────────────────────────────────
celery==5.6.2
kombu==5.6.2
amqp==5.3.1
billiard==4.2.4
vine==5.1.0
redis==7.1.1

# Click (Celery CLI deps)
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
prompt_toolkit==3.0.52
wcwidth==0.6.0

# ── Database ─────────────────────────────────────────────────────
psycopg2-binary==2.9.11
sqlalchemy>=2.0.0

# ── Data Science / ML ────────────────────────────────────────────
numpy==2.4.2
pandas==3.0.0
scipy==1.17.0
scikit-learn==1.8.0
joblib==1.5.3
threadpoolctl==3.6.0
matplotlib==3.10.8
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
pillow==12.1.0
pyparsing==3.3.2

# ── AI / Vector Search ───────────────────────────────────────────
openai==1.30.1
pgvector==0.3.0

# ── Web Scraping ─────────────────────────────────────────────────
beautifulsoup4==4.12.3
lxml==5.2.2

# ── AWS ──────────────────────────────────────────────────────────
boto3==1.34.40

# ── Production Server ────────────────────────────────────────────
gunicorn==25.0.3

# ── Utilities ────────────────────────────────────────────────────
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
structlog==24.1.0
six==1.17.0
packaging==26.0
tzlocal==5.3.1
colorama==0.4.6
